Big commit. Lots of changes everywhere.
I've added a log function to the view. I need it to debug the code when testing new
functions,

I've also changed the main loop, the model and the UI to work with in two steps when
a player performs an action. It's neccessary for the 'copy' function but is a nice
touch also for 'add' and 'remove'.